ai_meeting_assistant/requirements.txt
# Core Dependencies
streamlit>=1.28.0
pyaudio>=0.2.11
faster-whisper>=0.9.0
# torch>=2.0.0  # Install separately - see installation guide
# torchaudio>=2.0.0  # Install separately - see installation guide

# Speaker Identification
pyannote.audio>=3.1.0

# NLP and Text Processing
transformers>=4.30.0
spacy>=3.6.0
keybert>=0.8.0
sentence-transformers>=2.2.0

# Emotion Detection
datasets>=2.14.0

# API Integration
wikipedia>=1.4.0
requests>=2.31.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydub>=0.25.0
